+++ frontend/src/features/pressreleases/edit/components/press-release-edit-client.tsx
@@ -1,0 +1,85 @@
+"use client";
+import React, { useState, useCallback, useEffect } from "react";
+import { useDebouncedCallback } from "use-debounce";
+import markdownHtml from "zenn-markdown-html";
+
+import { useMarkdownEditor } from "@/features/pressreleases/edit/hooks/useMarkdownEditor";
+import { Preview } from "@/features/pressreleases/edit/components/preview";
+import Header from "@/features/pressreleases/edit/components/header";
+import LeftSidebar from "@/features/pressreleases/edit/components/leftSidebar";
+import RightSidebar from "@/features/pressreleases/edit/components/rightSidebar";
+
+export default function PressReleaseEditClient() {
+  const [html, setHtml] = useState<string>("");
+  const [isEdit, setIsEdit] = useState<boolean>(true);
+  const [canSave, setCanSave] = useState<boolean>(false);
+  const [article, setArticle] = useState({
+    title: "",
+    subtitle: "",
+    doc: "",
+  });
+
+  const savePreview = useCallback(() => {
+    const html = markdownHtml(article.title + "\n" + article.doc);
+    setHtml(html);
+    setCanSave(true);
+    console.log(article.title + "\n" + article.doc);
+  }, [article.doc, article.title]);
+
+  // 入力の最終操作から0.5秒後に自動保存
+  const debouncedAutoSave = useDebouncedCallback(() => {
+    const html = markdownHtml(article.title + "\n" + article.doc);
+    setHtml(html);
+  }, 0.5);
+
+  useEffect(() => {
+    debouncedAutoSave();
+    return () => {
+      debouncedAutoSave.cancel();
+    };
+  }, [article.doc, debouncedAutoSave]);
+
+  const { editor, navigateToLine, navigateToEnd } = useMarkdownEditor({
+    doc: article.doc,
+    setDoc: (doc) => setArticle({ ...article, doc }),
+    savePreview,
+  });
+
+  return (
+    <div className="flex flex-col h-screen">
+      <Header canSave={canSave} doc={article.doc} />
+      <div className="flex h-full flex-1">
+        <LeftSidebar doc={article.doc} onJump={(line) => navigateToLine(line)} />
+        {/* メインコンテンツエリア */}
+        <div className="h-full w-full relative">
+          {isEdit ? (
+            <div className="h-full bg-gray-50 flex flex-col">
+              <div className="mx-8 pt-10 flex-1">
+                <input
+                  className="w-full text-4xl font-bold tracking-wider focus:outline-none bg-transparent"
+                  placeholder="リリースタイトルを入力"
+                  value={article.title}
+                  onChange={(e) => setArticle({ ...article, title: e.target.value })}
+                />
+                <div ref={editor} className="mt-10" />
+              </div>
+              {/* クリック可能な下部余白 */}
+              <div
+                className="sticky bottom-0 h-8 bg-gray-50 hover:bg-gray-100 cursor-pointer flex items-center justify-center text-gray-400 hover:text-gray-600 border-t border-gray-200"
+                onClick={navigateToEnd}
+                title="クリックして最下行に移動"
+              >
+                <div className="text-sm">↓ クリックして最下行に移動 ↓</div>
+              </div>
+            </div>
+          ) : (
+            <div className="m-10">
+              <Preview html={html} />
+            </div>
+          )}
+        </div>
+        <RightSidebar doc={article.doc} isEdit={isEdit} onToggle={setIsEdit} />
+      </div>
+    </div>
+  );
+}
